+++ lib/blake3_ffi.dart
@@ -13,5 +13,5 @@
 library;
 
 export 'src/functions.dart'
-    show blake3, blake3DeriveKey, blake3Hex, blake3Keyed;
+    show blake3, blake3DeriveKey, blake3DeriveKeyHex, blake3Hex, blake3Keyed, blake3KeyedHex;
 export 'src/hasher.dart' show Blake3Hasher, blake3KeyLength, blake3OutLength;